Extract API and view endpoints to controller modules
