+++ backend/controllers/views.js
@@ -1,0 +1,12 @@
+const viewsRouter = require("express").Router();
+const mongo = require("../mongo");
+
+// === VIEWS ===
+viewsRouter.get("/info", async (_, response) => {
+    const persons = await mongo.fetchPersons();
+    const personCount = persons.length;
+    const timeNow = new Date();
+    response.render("info", { personCount, timeNow });
+});
+
+module.exports = viewsRouter;
